Optimization of Docker Container and it's Runtime Scripts for Deployment (#3)
The following commits (by title and is assumed to be impactful) are the ones that contribute to the development of this PR. 
This took 2.5 days since I was trying per case which leads to 95 commits. By this point, other changes that I can bring will be implemented on to the main branch. Workflows that require to refer the branch name will be changed later.

Anyway, I was able to optimize my docker container (and also from making it work), I was able to cache internally without using any VMs and other such stuff. I will be crediting the articles later aside from the ones that I dropped in the README that is not formatted.

* Experimental: Add Docker Cacher Action Workflow
* Update: Moved python-dotenv to Dev and Adjusted Code
* Fallback: Added Action Docker Layer Caching and other adjustments
* Adapt: Attempting To Use Caching Layers
* Adapted -> https://github.com/CodexLink/github-actions-docker-layer-caching-poc/blob/master/.github/workflows/dockerimage.yml
* Removed step "Move Cache"
* Added debug on setup-buildx-action
* ...: Let user/app be name of me/repo.
* Removed those redundant paths again
* Test to see for cached output
* Potential Post Candidate: Enabling Docker Test Again and Check for
Cache.
* Docker Optimization: Consolidation of RUN Commands + --no-cache-dir on
PIP
* Docker Optimization (try): One Liner RUN?
* Docker Optimization (reduce): Removed \ at the end
* Docker Script Deploy: Enabled with Cache Key github.sha
* Docker Optimization (explicit): Bump Poetry to 1.1.7 indefinitely.
* Docker Optimization (Seperate): Poetry Install from Other Instructions
* Docker Optimization (sepcific change): Deduced Path
* Every Commit Matters: Forgot chmod change path.
* File Inclusion: poetry.lock and pyproject.toml