Checkpoint: Introduced AsyncAPI, Typing, Exceptions, and Refactors
It's been a while since I took a break due to some medical issues. Here
are the changes that I have done so far. Keep note that these are
incomplete, this commit was a checkpoint because I have 3 async
functions contradicting of pre-requisites.

on .env_template:
+ Added GITHUB_API_URL, GITHUB_ACTOR and SHOW_HOURS_MINUTES_ELAPSED
  (will fix naming later)

on README.md:
+ Added 2 more stackoverflow references

on api.py:
+ Partially Implemented Async HTTP Request Class

on args.py:
+ Code Formatting with Black + Removed Else Clause

on badge.py:
- Removed establishing test connections code (The code was moved to api.py)

on client.py:
+ Added some log prints.

on entrypoint.py:
- Reinvented code steps in __start__
- Commented Extra Exception Handlers in __end__ (Subject to remove.)
/ Logs Adjusted.
/ asyncio_sleep from 0.5 to 0.8 (Subject to change.)
/ _check_dotenv and _init_logger was now converted to regular function
as a preparatory step.
- Removed <event_loop>.stop() to avoid displaying errors.

* test_api and test_readme_fetch is committed for future use. Though, they
were subjected for removal in the future commits.

on constants.py:
- Removed else clause before raise statement.
+ Introduced Enums RESTResponse and ResponseTypes

on exceptions.py:
+ Introduced potential exceptions: (GithubRateLimitedError,
  SessionStatusNotOkay, and SessionREquestHTTPError), subject for
  change though.

on typing.py:
+ Added new type, HttpsURL, subject for change though. The name looks
  incomplete.

It's been a while after 4 days.