Update: Removed PyGithub + Resolved Unused Imports and more!
For this commit, I was able to fix everything in the `prepare()`, as it
was able to run asynchronously. With the removal of PyGithub, I was also
able to implement basic functionality of making requests with Session,
so everything is fine. It's time to handle Discord and BadgeConstructor
then.

Changes:
on .env_template:
/ Fixed `INPUT_ITHUB_API_URL` to `SHOW_HOURS_MINUTES_ELAPSED` (to the next
commit)
/ Changed `INPUT_PROFILE_REPOSITORY` to `INPUT_PROFILE_REPO`

on poetry.lock: Updated.

on pyproject.toml:
- Removed DevPackages | PyGithub, pytest, PyYAML

on api.py:
+ Resolved Unused Imports
+ Added sleep() and re-arranged logging.
+ Finished test_api_conn function.
+ Initialized github_action_repo() | with FERCH_README finished.
/ Resolved Invalid ENUM from RESP_STATUS to RESPONSE_STATUS.

on args.py:
/ Import Only asyncio.gather().
+ Gathered super().__ainit__ with self.__load_args.
- Removed await asyncio_sleep(0) on end.

on entrypoint.py: (more changes will be reflected on next commit)
+ Added GithubRunnerActions ENUM
+ Gathered self.prepare() and __ainit__().
+ Moved Condition self.args_container.running_on_local + await
  asyncio_sleep(0.001) to give time for ArgsContainer to Instantiate.
+ Removed PyGithub Related Functions and Replaced with Async API
  Functions.
+ Gathered self.github_api_connect() and self.github_action_repo()
+ Added float by .2f in string.
/ Adjusted asyncio_sleep from 0.8 to 0.4.
+ Removed conditional checks, moved from function calling.

on test_readme_fetch.py: Updated.

on constants.py:
+ Added GITHUB_API_URL, GITHUB_ACTOR to be handled in `resolved_envs`.
/ Renamed `INPUT_PROFILE_REPOSITORY` to `INPUT_PROFILE_REPO`

on typing.py:
+ Added Base64 under str class.