Update: Added Badge Validator + Enumerations and Typing and More.
This commit constsi of a day work that I have for a while. I was unable
to commit yesterday because I was doing something else. Can't focus that
much, but I was able to do something impacful to this commit. See the
following changelogs below.

Changelogs:

on README.md: Added New References

on api.py:
- Use Approach of Refer to Variables to Process (I don't want this to be
  a magic game where I have to use my time to look for some variables
  that is involved to a certain process) [see entrypoint.py:179 for this
  commit.]
+ Added More Logging.
/ Changed function name from `github_action_repo` to `exec_api_actions`.
/ Make `exec_api_actions` return values instead of holding it in the
`Class`.

on args.py: Changed argument for running local from `--running-on-local` to `--local`. (Abbreviations were changed as well.)

on badge.py:
+ Added functions for handling base64 strings (`_handle_b64`), identifying of badge in README (`checkbadge_identifier`) and instantiated function for preparing the badge during other async work. (`prepare_badge_metadata`)
- Removed `__ainit__` due to missing variables.

on client.py: Changed some logging information.

on entrypoint.py:
+ Encapsulated Environment Sterilization to `_resolve_envs()` from
  `prepare()`
/ Let function with argument `FETCH_README` to be executed inside
`process()` instead of `prepare()` due to inability to await.

on test_readme_fetch.py (Temporary): Added part to identify badge in
README.

on constants.py:
/ Changed Regex Formula Group (`Identifier`) to match other symbols like
`()-` aside from what is included in `\w`.
+ Added Base64Actions Enum and B64_ACTION_FILENAME for writing README.
- Removed DISCORD_ON_READY_MSG.

on typing.py:
+ Added Types (ResolvedHTTPResponse and BadgeStructure)

Almost there and we are about to finish this stupid pet project that I
thought I can finish for about 3 weeks. ^^