Update (codebase): Properly Constructed Badge Strings
This commit contains 2 days worth of work. I was unable to commit via
checkpoint because it was convicing for me to say that I should commit
the one that is in the right state. So for this commit, it just that. We
are nearing in completion.

Some functionalities in todo might be ignored at some point till I hit
the post-development. I just wanna get this done to enjoy myself with
Full-Stack Developer Tutorials because I'm nearing graduation.

Here are the changelogs so far:

on .env_template, action.yml, and README.md: Updated Parameters along
with Information.

on args.py: Added missing import library `os`.

badge.py:
/ Added `return` at function `_handle_b64` -> Union[Any, None]
+ Constructed Inline Conditional Hell for Constructing Badge Strings.
  (This one took me a day to finish, but only the initial. More to come
  from this one.)

on client.py:
+ Added debugger output for `Is WS Rate-Limited?` Output is
  uninformative btw.
/ Reflect Changes from `Enum` changes that is factored by 3 files
indicated on the first element.

on entrypoint.py:
/ Added mechanism resolver specifically for `Enum` by UPPERCASING.
+ Added `ContextOnSubject` on `_enum_candidates` for Iteration.
/ TEMPORARILY Removed Script Runtime Limitation. To be enabled soon.

on constants.py:
/ Appended `_ACTIVITY` on every elements from
`PreferredActivityDisplay`.
+ Resolved `DISABLED` from two enums (`PreferredTimeDisplay` and
  `ContextOnSubject`)
+ Added BADGE_BASE_SUBJECT for Basic No Conditions Met in Badge.
/ Changed parameter from `APPEND_DETAIL_PRESENCE` to be merged under
`PREFERRED_PRESENCE_CONTEXT`
+ Added `TIME_DISPLAY_SHORTHAND` and `STATUS_CONTEXT_SEPERATOR`
  parameters.
/ From DISCORD_STRUCT_BLUEPRINT, key `presence` has been changed to
`activities`.
+ `TIME_STRINGS` has been added in the constants.

By this point, I was trying my best to ignore everything that I can
improve at this point. I just wanna finish it first before I tried it.

The next commit should introduce full functionality of mechanism from
constructing the strings to embedding it to the badge. Once that's done,
time for Online Test. But before that, I should be attempting to merge
it in the `main`. And then create a PR to create a new branch `v1`.