Update (Append Logic): Finished Condition Construction. Wrap up to Badge Left.
This commit is focused on the construction of logic for the badge. It
was done, but wrapping it up into a full badge is not yet initialized. Yesterday's
workload before this commit, was too light but took a while because of
confusion.

Next commit asserts and ensures that the badge is completed and ready
for the README modificaition or onwards (depending on how focused I
would be in that particular day).

3-4~ elements left before finalization. Estimated to finish is until
Saturday.