Update (optimization): Removed Redundancy + Refactor
This commit contains a lot of reworks and stuff and is contained for
atleast 2.5 days. The reason why I held back because I have a lot of
things to consider, such as the design. Instead of finishing it, I just
want to take a look further of what I can really do to optimize the
script. There has been a lot of features that has been removed because
of how stupid it is.

Anyway, this is the changelog for this commit (it's quite long):

for all test_xxx.py: Moved to /test/ folder. Contents Update based on
testing.

on README.md:
+ Added StackOverflow Links (Unordered)

on api.py:
/ Renamed from `AsyncRequestAPI` to `AsyncGithubAPI`
- Removed Unecessary BS.
- Removed `test_api_conn`
/ `exec_api_actions` return types has been fixed.
+ Invoke auth parameters on request when `self.is_auth` not True.
/ `_request` return types has been fixed.
- `_reopen_conn()` is removed because it was unused.

on args.py: Removed, components move to `utils.py` as a function
instead.

on badge.py:
+ Added placeholder function `_replace_content()`
/ Improved Condition in `check_and_update_badge()` (before was
`check_badge_identifier()`)
/ String Concatenation Logic Improved.

on entrypoint.py:
- Removed redundant functions: `prepare()` and `process()`.
- Moved utility functions to `utils.py`
/ Added Anti-Redundant (Length and Content Sets) Tasks Counter Display.
/ use `gather()` for closing connection sessions of aiohttp and discord.

Keep in mind that, the expected delivery feature promise has been
broken. Will try to deliver that certain commit that delivers the
certain thing that finishes this pet project.

on utils.py: New! Wrapped all utility functions as a class.

on constants.py:
/ Rearranged Badge Generator Constants.
+ Rearranged and added more enumerations such as `LoggerLevelCoverage`
  and `ExitReturnCodes`.

on typing.py:
+ Rearranged and added types (`ResolvedHTTPResponse` ->
  `ResolvedClientResponse`), `BadgeElements`, `ClientResponseOK`, and `ClientResponseStatus`.

I have to recheck the flow and attempt the previous implementation of
`as_completed` to see if there's an impact. And then I will do the
commit now. This ensures that after post-development, there are less
things to deal with, because I wanted to learn Web Development this
time.