Update (codebase by 95%): Tasks Finalized + Runtime 4.5s Less
This commit updates the mainstream `dev` branch to acommodate with my
inactivity. I was unable to work for 3 days for some reason, but was
able to work for 2 days but lightly. This commit ensures that everything
is running and can sub 4 seconds. The only left from this project is
basically the commit.

ANyway, here are the changes:

on README.md: Added more stackoverflow links.

on api.py:
/ Changed code structure for accomodating the functionality for
GithubRunnerActions.COMMIT_CHANGES.
/ on `_request`, prepped every possible parameters (except for commits
since this was related to `headers`.
/ Added dynamic attribute method calling with `getattr()`.

on badge.py:
/ on `_handle_b64`, implies direct or inline function and NewType
calling to be returned immediately. (This works on both ENCODE and
DECODE)
- Removed File Write and Update, goes beyond in-memory modification of
  string.
/ Allow `await` of other tasks declared in `entrypoint.py`
/ Direct return from calling `_handle_b64` for final output.
/ `construct_badge()` has typed return of `BadgeStructure`
/ Updated some logging strings.
+ `construct_badge()` now returns final outptu of the badge
  construction.

on client.py: Updated some logging output strings.

on entrypoint.py: Order of Tasks with `create_tasks()` was now
finalized.

on constants.py:
/ Changed Base64Actions Enum of `ENCODE_B64_TO_FILE` to
`ENCODE_BUFFER_TO_B64`
! `DECODE_B64_TO_FILE` will be changed to `DECODE_B64_TO_BUFFER` on the
next commit!
- Removed some unused enums on `GithubRunnerActions` (deduced to two.)

on typing.py:
+ Added new types: `Base64Bytes`, `Base64String`, `READMERawContent`,
  and `READMEContent`.
- Removed generalized type: `Base64`

on readme_fetch_to_read.py: Added some print functions and invoked auth
parameter.

General Updates:
- Some of the functions are now **typed**!
- Some resolved but unused imports are retained for the purpose of using
  them in post-development or candidate commit state of the `dev`
  branch.

The reason why I have to commit this one because the structure that I
need to build is too long and may require more a day or so to test it.
Keep in mind that, my special repository may be affected so that I can
live test it. This ensures that I can actually deliver it `v1` at some
point in time.

Expected finish of the development is **within this week**.