Update + Clean-up (Codebase): Annotated and Fixed.
This commit updates all files with several fixes. Most of the commit is
focused on annotation so that when I visit this project in the future, I
can read it without having cancer.

There are few more commits left before I focus on other stuff and mostly
work in `cutting-edge` branch.