Update (Handler): Added Github Actions Message Output Compatibility.
For this commit, there are quite some changes that I have done so that
things are going to be fine in the post-development.

So far I implemented outputting of data, and they, I also implemented
the abiltiy to compare the old readme with the new changes, this was
done so that we have avoid spamming commits when there are no changes,
that would also avoid 409 Conflict as well.

This commit may seem to be the last changes (hopefully) in the codebase.
The next possible commit would probably focused on README. Hopefully.

No promise btw.